+++ toro/src/main/java/im/ene/lab/toro/AbsVideoViewHolder.java
@@ -111,16 +111,6 @@ public abstract class AbsVideoViewHolder extends ToroViewHolder {
     return 0;
   }
 
-  @Override public float visibleAreaOffset() {
-    Rect videoRect = getVideoRect();
-    Rect parentRect = getRecyclerViewRect();
-    if (parentRect != null && !parentRect.contains(videoRect) && !parentRect.intersect(videoRect)) {
-      return 0.f;
-    }
-
-    return mVideoView.getHeight() <= 0 ? 1.f : videoRect.height() / (float) mVideoView.getHeight();
-  }
-
   @Override public boolean wantsToPlay() {
     // Default implementation
     return visibleAreaOffset() >= 0.75;
@@ -134,8 +124,9 @@ public abstract class AbsVideoViewHolder extends ToroViewHolder {
     mPlayable = true;
   }
 
-  @Override public void onPlaybackError(MediaPlayer mp, int what, int extra) {
+  @Override public boolean onPlaybackError(MediaPlayer mp, int what, int extra) {
     mPlayable = false;
+    return super.onPlaybackError(mp, what, extra);
   }
 
   @NonNull @Override public View getVideoView() {