+++ toro/src/main/java/im/ene/lab/toro/ToroViewHolder.java
@@ -21,6 +21,7 @@ import android.graphics.Rect;
 import android.media.MediaPlayer;
 import android.support.annotation.CallSuper;
 import android.support.annotation.Nullable;
+import android.util.Log;
 import android.view.View;
 
 /**
@@ -28,19 +29,19 @@ import android.view.View;
  */
 public abstract class ToroViewHolder extends ToroAdapter.ViewHolder implements ToroPlayer {
 
-  @Nullable private final VideoViewHolderHelper mHelper;
+  private final VideoViewItemHelper mHelper;
 
   private View.OnLongClickListener mLongClickListener;
 
   public ToroViewHolder(View itemView) {
     super(itemView);
-    mHelper = Toro.getHelper(this);
+    mHelper = Toro.RECYCLER_VIEW_HELPER;
     if (allowLongPressSupport()) {
       if (mLongClickListener == null) {
         mLongClickListener = new View.OnLongClickListener() {
           @Override public boolean onLongClick(View v) {
-            return mHelper != null && mHelper.onItemLongClick(ToroViewHolder.this,
-                ToroViewHolder.this.itemView, ToroViewHolder.this.itemView.getParent());
+            return mHelper.onItemLongClick(ToroViewHolder.this, ToroViewHolder.this.itemView,
+                ToroViewHolder.this.itemView.getParent());
           }
         };
       }
@@ -63,8 +64,7 @@ public abstract class ToroViewHolder extends ToroAdapter.ViewHolder implements T
       if (mLongClickListener == null) {
         mLongClickListener = new View.OnLongClickListener() {
           @Override public boolean onLongClick(View v) {
-            return mHelper != null && mHelper.onItemLongClick(ToroViewHolder.this, itemView,
-                itemView.getParent());
+            return mHelper.onItemLongClick(ToroViewHolder.this, itemView, itemView.getParent());
           }
         };
       }
@@ -74,10 +74,12 @@ public abstract class ToroViewHolder extends ToroAdapter.ViewHolder implements T
 
     super.setOnItemLongClickListener(new View.OnLongClickListener() {
       @Override public boolean onLongClick(View v) {
+        boolean longClicked = false;
+
         if (mLongClickListener != null) {
-          mLongClickListener.onLongClick(v);  // we can ignore this boolean result
+          longClicked = mLongClickListener.onLongClick(v);  // we can ignore this boolean result
         }
-        return listener.onLongClick(v);
+        return listener.onLongClick(v) && longClicked;
       }
     });
   }
@@ -89,25 +91,19 @@ public abstract class ToroViewHolder extends ToroAdapter.ViewHolder implements T
 
   @CallSuper @Override public void onAttachedToParent() {
     super.onAttachedToParent();
-    if (mHelper != null) {
-      mHelper.onAttachedToParent(this, itemView, itemView.getParent());
-    }
+    mHelper.onAttachedToParent(this, itemView, itemView.getParent());
   }
 
   @CallSuper @Override public void onDetachedFromParent() {
     super.onDetachedFromParent();
-    if (mHelper != null) {
-      mHelper.onDetachedFromParent(this, itemView, itemView.getParent());
-    }
+    mHelper.onDetachedFromParent(this, itemView, itemView.getParent());
   }
 
   /**
    * Implement from {@link MediaPlayer.OnPreparedListener#onPrepared(MediaPlayer)}
    */
   @CallSuper @Override public final void onPrepared(MediaPlayer mp) {
-    if (mHelper != null) {
-      mHelper.onPrepared(this, itemView, itemView.getParent(), mp);
-    }
+    mHelper.onPrepared(this, itemView, itemView.getParent(), mp);
   }
 
   /**
@@ -116,9 +112,7 @@ public abstract class ToroViewHolder extends ToroAdapter.ViewHolder implements T
    * should use {@link ToroPlayer#onPlaybackStopped()}
    */
   @Override public final void onCompletion(MediaPlayer mp) {
-    if (mHelper != null) {
-      mHelper.onCompletion(this, mp);
-    }
+    mHelper.onCompletion(this, mp);
   }
 
   /**
@@ -127,23 +121,21 @@ public abstract class ToroViewHolder extends ToroAdapter.ViewHolder implements T
    * should use {@link ToroPlayer#onPlaybackError(MediaPlayer, int, int)}
    */
   @Override public final boolean onError(MediaPlayer mp, int what, int extra) {
-    return mHelper != null && mHelper.onError(this, mp, what, extra);
+    return mHelper.onError(this, mp, what, extra);
   }
 
   /**
    * Implement from {@link MediaPlayer.OnInfoListener#onInfo(MediaPlayer, int, int)}
    */
   @CallSuper @Override public final boolean onInfo(MediaPlayer mp, int what, int extra) {
-    return mHelper != null && mHelper.onInfo(this, mp, what, extra);
+    return mHelper.onInfo(this, mp, what, extra);
   }
 
   /**
    * Implement from {@link MediaPlayer.OnSeekCompleteListener#onSeekComplete(MediaPlayer)}
    */
   @CallSuper @Override public final void onSeekComplete(MediaPlayer mp) {
-    if (mHelper != null) {
-      mHelper.onSeekComplete(this, mp);
-    }
+    mHelper.onSeekComplete(this, mp);
   }
 
   @Override public int getPlayOrder() {
@@ -169,8 +161,8 @@ public abstract class ToroViewHolder extends ToroAdapter.ViewHolder implements T
 
   }
 
-  @Override public void onPlaybackError(MediaPlayer mp, int what, int extra) {
-
+  @Override public boolean onPlaybackError(MediaPlayer mp, int what, int extra) {
+    return true;  // don't want to see the annoying dialog
   }
 
   @Override public void onPlaybackInfo(MediaPlayer mp, int what, int extra) {
@@ -181,30 +173,38 @@ public abstract class ToroViewHolder extends ToroAdapter.ViewHolder implements T
 
   }
 
+  private static final String TAG = "ToroViewHolder";
+
   @Override public float visibleAreaOffset() {
     Rect videoRect = getVideoRect();
     Rect parentRect = getRecyclerViewRect();
-    if (parentRect != null && !parentRect.contains(videoRect) && !parentRect.intersect(videoRect)) {
+
+    if (parentRect != null && (parentRect.contains(videoRect) || parentRect.intersect(videoRect))) {
+      float visibleArea = videoRect.height() * videoRect.width();
+      float viewArea = getVideoView().getWidth() * getVideoView().getHeight();
+      return viewArea <= 0.f ? 1.f : visibleArea / viewArea;
+    } else {
       return 0.f;
     }
-
-    return getVideoView().getHeight() <= 0 ? 1.f
-        : videoRect.height() / (float) getVideoView().getHeight();
   }
 
   protected final Rect getVideoRect() {
     Rect rect = new Rect();
-    getVideoView().getGlobalVisibleRect(rect, new Point());
+    Point offset = new Point();
+    getVideoView().getGlobalVisibleRect(rect, offset);
+    Log.i(TAG, "VideoView Rect: " + getVideoId() + " | " + offset + " | " + rect);
     return rect;
   }
 
-  protected final Rect getRecyclerViewRect() {
-    if (itemView.getParent() == null) {
+  @Nullable protected final Rect getRecyclerViewRect() {
+    if (itemView.getParent() == null) { // view is not attached to RecyclerView
       return null;
     }
 
     Rect rect = new Rect();
-    ((View) itemView.getParent()).getGlobalVisibleRect(rect, new Point());
+    Point offset = new Point();
+    ((View) itemView.getParent()).getGlobalVisibleRect(rect, offset);
+    Log.e(TAG, "Parent    Rect: " + getVideoId() + " | " + offset + " | " + rect);
     return rect;
   }
 
@@ -230,5 +230,14 @@ public abstract class ToroViewHolder extends ToroAdapter.ViewHolder implements T
 
   @Override public int getAudioSessionId() {
     return 0;
+  }
+
+  /**
+   * Indicate that this Player is able to replay right after it stops (loop-able) or not.
+   *
+   * @return true if this Player is loop-able, false otherwise
+   */
+  @Override public boolean isLoopAble() {
+    return false;
   }
 }